Changed behaviour of "eachComplex" for the WoodstoxXmlSerializer.
The container name provided to "eachComplex" will now be used as a container for each object
(for XML) instead of a container for the collection (which is still the case for json).
This makes it possible to serialize objects without a "surrounding" container which will make
json collections more compact/correct.

pom update:
com.natpryce.maybe dependency is now expected to be found in the  Maven repo. Install locally
or in a centralized repository.